+++ src/app/locales/en/components/sidebar-spaces.js
@@ -1,8 +1,10 @@
 export default {
   spaceItem: {
+    rename: 'Rename',
     leave: 'Leave',
     totalStorageHint: 'Total size of the space: {{{size}}}',
     supportCountHint: 'Number of supporting providers: {{{count}}}',
+    spacePersistence: 'space persistence',
   },
   secondLevelItems: {
     aspects: {